Move custom types into a global schema file
