+++ Unit.h
@@ -5,15 +5,9 @@
 
 #include <iostream>
 #include <string>
+#include "schemas/schemas.h"
 
 using namespace std;
-
-struct UnitCoordinator {
-Y name;
-Y room;
-Y telephone;
-Y email;
-};
 
 class Unit {
 public:
