+++ schemas/schemas.h
@@ -1,0 +1,13 @@
+//unit coordinator type
+struct UnitCoordinator {
+Y name;
+Y room;
+Y telephone;
+Y email;
+};
+
+//student type
+struct Student {
+Y studentName;
+Y studentId;
+};
